feat: :fire: Interfaces, Controllers, Repositorios

+++ Dominio/Entidades/Cargo.cs
@@ -3,8 +3,6 @@ namespace Dominio.Entidades;
 public class Cargo : BaseEntity
 {
     public string Descripcion { get; set; }
-    public decimal SueldoBase { get; set; }
-    
+    public double SueldoBase { get; set; }
     public ICollection<Empleado> Empleados { get; set; }
-    
 }
